Homework 5 part 1 done

+++ project/bot/src/main/java/ru/tinkoff/edu/java/bot/model/dto/request/LinkUpdateRequest.java
@@ -3,7 +3,7 @@ package ru.tinkoff.edu.java.bot.model.dto.request;
 import java.net.URI;
 import java.util.List;
 
-public record LinkUpdateRequest(Integer linkId,
+public record LinkUpdateRequest(Long linkId,
                                 URI url,
                                 String description,
-                                List<Integer> tgChatIds) {}
+                                List<Long> tgChatIds) {}
